movie_recommender/requirements.txt
pandas>=2.1,<3
numpy>=1.26,<3
scikit-learn>=1.4,<2
scipy>=1.11,<2
fastapi>=0.115
uvicorn[standard]>=0.30
scikit-surprise==1.1.4
joblib>=1.3,<2
